requirements.txt
cx-Freeze==6.9
cx-Logging==3.0
cycler==0.11.0
fonttools==4.28.5
importlib-metadata==4.10.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.0
packaging==21.3
Pillow==9.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
scipy==1.7.3
six==1.16.0
yattag==1.14.0
zipp==3.7.0
